+++ src/app.tsx
@@ -4,7 +4,9 @@ import { Router } from 'src/routes/sections';
 import { useScrollToTop } from 'src/hooks/use-scroll-to-top';
 import { ThemeProvider } from 'src/theme/theme-provider';
 import { Iconify } from 'src/components/iconify';
+import { SnackbarProvider } from 'notistack';
 import { Provider } from 'react-redux';
+import { ToastProvider } from './components/snackBar/ToastContext';
 import store from './store/store';
 
 // ----------------------------------------------------------------------
@@ -35,7 +37,11 @@ export default function App() {
   return (
     <Provider store={store}>
       <ThemeProvider>
-        <Router />
+        <SnackbarProvider maxSnack={3}>
+          <ToastProvider>
+            <Router />
+          </ToastProvider>
+        </SnackbarProvider>
       </ThemeProvider>
     </Provider>
   );
