+++ src/interface/UserHistory.ts
@@ -1,0 +1,8 @@
+export interface History {
+  id: string;
+  source: string;
+  destination: string;
+  rating: number;
+  distance: string;  // New field
+  timeTaken: string; // New field
+}
