+++ src/layouts/simple/layout.tsx
@@ -5,7 +5,7 @@ import Alert from '@mui/material/Alert';
 
 import { RouterLink } from 'src/routes/components';
 
-import  Logo  from '../../../public/assets/images/main-logo.png';
+import Logo from '../../../public/assets/images/main-logo.png';
 
 import { Main, CompactContent } from './main';
 import { LayoutSection } from '../core/layout-section';
@@ -43,11 +43,7 @@ export function SimpleLayout({ sx, children, header, content }: SimpleLayoutProp
                 This is an info Alert.
               </Alert>
             ),
-            leftArea: (<img
-              src={Logo}
-              alt="Logo"
-              className="h-[50px] w-[50px]"
-            />),
+            leftArea: <img src={Logo} alt="Logo" className="h-[3.125rem] w-[3.125rem]" />,
             rightArea: (
               <Link
                 href="#"
